Backend/requirements.txt
asgiref==3.7.2
django==4.2.5
Pillow>=11.0.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
phonenumbers==8.13.7
django-environ==0.11.2
psycopg2-binary
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
celery-progress==0.3
django-celery-beat==2.5.0
flower==2.0.1
django-rest-passwordreset==1.3.0
django-redis==5.2.0
celery-singleton==0.3.1
django-cachalot==2.6.2